add textureatlas class and change logic in generate_texture_atlas()
There is a bug using the scale var in _atlas_single_channel because it is undefined. The other new code has not been tested yet, but the design is there. If you want to work on this be aware that you may have to work through a lot of bugs before it works. Sorry!